Add Jupyter Notebook support.
# Usage
Put the symbol `# +!+=<name>` anywhere in the Jypyter Code Block. The checker will be injected at that spot.

Each input file is a python checker code.

    def ___cafe_jupyter__checker___<name>():
        print(test_func(5,2))
        print(test_func(9,4))
        sys.exit(0)

The `sys.exit(0)` is required to ensure the checker are actually executed and following codes are not run.

If the student deleted the `# +!+=<name>`  marker from the notebook, the code execution will fall through to the end of file and will result in a x verdict.

The solution file is the expected output from the checker code.

It uses the same venv as regular Python to run the code.

# Technical detail

During compilation step, ţhe submitted .ipynb is processed to extract only the code part. The checker marker is transformed into:

    if '___cafe_jupyter__checker___<name>' in dir(): ___cafe_jupyter__checker___name()

This will execute the checker function if exists.

It will also prepend `import sys` and `sys.exit(1)` at the end of file.

The compilation step then generate a run script, which will prepend each input to the processed submission code, then run the code as normal (but without input redirect).